fix: move loading logic to peers table
